Ongoing test branch for core integration stuff.
Basically, this branch is testing functional changes in the plugin that
interact with corresponding a11y changes in core Blockly in order to
have a working screenreader a11y experience in the plugin's sample.

This branch will not be merged in and is only meant for sharing context
of this work as it is ongoing.